requirements.txt
pyserial>=3.5
protobuf>=7.34.0
